[branch_course] Configured token provider and jwt filter for notification microservices

+++ authuser/src/main/java/com/ead/authuser/services/UserService.java
@@ -17,16 +17,18 @@ import com.ead.authuser.repositories.RoleRepository;
 import com.ead.authuser.repositories.UserRepository;
 import com.ead.authuser.responses.ImageResponse;
 import com.ead.authuser.responses.TokenResponse;
+import com.ead.authuser.security.AuthorizationCurrentUserService;
 import com.ead.authuser.security.JwtProvider;
+import com.ead.authuser.security.UserDetailsImpl;
 import com.ead.authuser.utils.UserUtils;
 import com.ead.authuser.utils.ValidaCPF;
-import com.google.common.collect.ImmutableSet;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.BeanUtils;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -37,7 +39,8 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.time.ZoneId;
-import java.util.List;
+import java.util.HashSet;
+import java.util.Set;
 import java.util.UUID;
 
 import static com.ead.authuser.constants.UserMessagesConstants.*;
@@ -54,23 +57,29 @@ public class UserService {
     private final PasswordEncoder passwordEncoder;
     private final JwtProvider jwtProvider;
     private final AuthenticationManager authenticationManager;
+    private final AuthorizationCurrentUserService authorizationCurrentUserService;
 
     public Page<UserDto> findAllUsers(Specification<User> spec, Pageable pageable) {
         return userUtils.toListUserDto(userRepository.findAll(spec, pageable));
     }
 
     public User findUserById(UUID id) {
-        return userRepository.findById(id)
-                .orElseThrow(() -> new UserNotFoundException(id));
+        var currentUserId = authorizationCurrentUserService.getCurrentUser().getId();
+        if (currentUserId.equals(id)) {
+            return userRepository.findById(id)
+                    .orElseThrow(() -> new UserNotFoundException(id));
+        } else {
+            throw new AccessDeniedException(authorizationCurrentUserService.getCurrentUser().getFullName() + " doesn't have the correct authorities to perform this action");
+        }
     }
 
-    public Page<CourseDto> getCoursesByUser(UUID userId, Pageable pageable) {
+    public Page<CourseDto> getCoursesByUser(UUID userId, Pageable pageable, String token) {
         return courseClientFeign.getAllCoursesByUser(CourseClientParams.builder()
                 .userId(userId)
                 .page(pageable.getPageNumber())
                 .size(pageable.getPageSize())
                 .sort(pageable.getSort().toString().replaceAll(": ", ","))
-                .build());
+                .build(), token);
     }
 
     public TokenResponse signIn(SignInDto request) throws BadCredentialsException {
@@ -91,7 +100,7 @@ public class UserService {
             throw new UserException(USUARIO_BLOQUEADO_MENSAGEM);
         }
 
-        user.getRoles().add(verifyRole(Roles.ROLE_INSTRUCTOR));
+        user.addRole(verifyRole(Roles.ROLE_INSTRUCTOR));
         user.setUserType(UserType.INSTRUCTOR);
         user.setLastUpdateDate(LocalDateTime.now(ZoneId.of("UTC")));
         userEventPublisher.publishUserEvent(userUtils.toUserEventDto(user), type);
@@ -100,9 +109,10 @@ public class UserService {
 
     @Transactional
     public UserDto saveUser(UserDto userDto, ActionType type) {
-        User user;
+        var userRegistered = new User();
 
         if (userDto.getId() == null) {
+
             if (!ValidaCPF.isCPF(userDto.getCpf())) {
                 throw new UserException(CPF_MENSAGEM + userDto.getCpf());
             }
@@ -116,61 +126,93 @@ public class UserService {
                 throw new UserException(USUARIO_EMAIL_EXISTENTE_MENSAGEM + userDto.getEmail());
             }
 
-            var roles = ImmutableSet.copyOf(List.of(verifyRole(Roles.ROLE_STUDENT)));
-
             userDto.setPassword(passwordEncoder.encode(userDto.getPassword()));
 
-            user = new User();
+            var user = new User();
 
             BeanUtils.copyProperties(userDto, user);
 
             user.setUserStatus(UserStatus.ACTIVE);
             user.setUserType(UserType.STUDENT);
-            user.setRoles(roles);
+            user.addRole(verifyRole(Roles.ROLE_STUDENT));
             user.setCreationDate(LocalDateTime.now(ZoneId.of("UTC")));
             user.setCurrentPasswordDate(LocalDateTime.now(ZoneId.of("UTC")));
-            log.debug("Method saveUser user created {} ", user.toString());
+            log.debug("Method saveUser user created {} ", user);
+
+            user.setLastUpdateDate(LocalDateTime.now(ZoneId.of("UTC")));
+
+            userRegistered = userRepository.save(user);
 
         } else {
-            user = userRepository.findById(userDto.getId()).orElseThrow(() -> new UserNotFoundException(userDto.getId()));
+            UserDetailsImpl currentLoggedUser = (UserDetailsImpl) authorizationCurrentUserService.getAuthentication().getPrincipal();
 
-            user.setUsername(userDto.getUsername());
-            user.setEmail(userDto.getEmail());
-            user.setFullName(userDto.getFullName());
-            user.setPhoneNumber(userDto.getPhoneNumber());
-            user.setUserStatus(UserStatus.valueOf(userDto.getStatus()));
-            user.setUserType(UserType.valueOf(userDto.getType()));
-            log.debug("Method saveUser user updated {} ", user.toString());
+            currentLoggedUser.getAuthorities().forEach(auth -> {
+                if (auth.getAuthority().equals(Roles.ROLE_ADMIN.name()) || auth.getAuthority().equals(Roles.ROLE_INSTRUCTOR.name())) {
+                    userDto.setAuthorized(true);
+                }
+            });
+
+            if (currentLoggedUser.getId().equals(userDto.getId()) || userDto.isAuthorized()) {
+                var user = userRepository.findById(userDto.getId()).orElseThrow(() -> new UserNotFoundException(userDto.getId()));
+
+                user.setUsername(userDto.getUsername());
+                user.setEmail(userDto.getEmail());
+                user.setFullName(userDto.getFullName());
+                user.setPhoneNumber(userDto.getPhoneNumber());
+                user.setUserType(UserType.valueOf(userDto.getType()));
+                log.debug("Method saveUser user updated {} ", user.toString());
+
+                userRegistered = user;
+            } else {
+                throw new AccessDeniedException(userDto.getId().toString());
+            }
         }
 
-        user.setLastUpdateDate(LocalDateTime.now(ZoneId.of("UTC")));
-
-        var userRegistered = userRepository.save(user);
-
-        userEventPublisher.publishUserEvent(userUtils.toUserEventDto(userRegistered), type);
+        publishEventUser(userUtils.toUserEventDto(userRegistered), type);
 
         return userUtils.toUserDto(userRegistered);
     }
 
     @Transactional
     public ImageResponse updateImage(UserDto userDto, ActionType type) {
-        var user = userRepository.findById(userDto.getId()).orElseThrow(() -> new UserNotFoundException(userDto.getId()));
+        UserDetailsImpl currentLoggedUser = (UserDetailsImpl) authorizationCurrentUserService.getAuthentication().getPrincipal();
 
-        user.setImageUrl(userDto.getImageUrl());
-        user.setLastUpdateDate(LocalDateTime.now(ZoneId.of("UTC")));
-        userEventPublisher.publishUserEvent(userUtils.toUserEventDto(user), type);
+        currentLoggedUser.getAuthorities().forEach(auth -> {
+            if (auth.getAuthority().equals(Roles.ROLE_ADMIN.name()) || auth.getAuthority().equals(Roles.ROLE_INSTRUCTOR.name())) {
+                userDto.setAuthorized(true);
+            }
+        });
 
-        return userUtils.toImageResponse(user);
+        if (currentLoggedUser.getId().equals(userDto.getId()) || userDto.isAuthorized()) {
+            var user = userRepository.findById(userDto.getId()).orElseThrow(() -> new UserNotFoundException(userDto.getId()));
+
+            user.setImageUrl(userDto.getImageUrl());
+            user.setLastUpdateDate(LocalDateTime.now(ZoneId.of("UTC")));
+            userEventPublisher.publishUserEvent(userUtils.toUserEventDto(user), type);
+
+            return userUtils.toImageResponse(user);
+        } else {
+            throw new AccessDeniedException(userDto.getId().toString());
+        }
     }
 
     @Transactional
     public void deleteUser(UUID id) {
-        var user = findUserById(id);
-        userRepository.delete(user);
-        userEventPublisher.publishUserEvent(userUtils.toUserEventDto(user), ActionType.DELETE);
+        var currentUserId = authorizationCurrentUserService.getCurrentUser().getId();
+        if (currentUserId.equals(id)) {
+            var user = findUserById(id);
+            userRepository.delete(user);
+            userEventPublisher.publishUserEvent(userUtils.toUserEventDto(user), ActionType.DELETE);
+        } else {
+            throw new AccessDeniedException(id.toString());
+        }
     }
 
     private Role verifyRole(Roles name) {
         return roleRepository.findByRoleName(name).orElseThrow(RoleNotFoundException::new);
+    }
+
+    private void publishEventUser(UserEventDto userEventDto, ActionType type) {
+        userEventPublisher.publishUserEvent(userEventDto, type);
     }
 }
